Create the Schema Converter library
We decided to create this library to expose the contract of specific objects and
allowing creating those objects via REST API.

Objects can be quite complex, and to instantiate them via an API request we need
to use the PHP's introspection API, unless we use PHP's serialization mechanism.
But still, exposing the contract of a class to create the object would require
quite some work.

Alternatively, we could have added some extra information to each class (like
PHP attributes) to make it easier to create them, but that would mean more work
for the teams when they need to create each class.

We decided to build this in-house because there isn't a library out there that
would expose the class contract in a user-friendly way—or in a way friendly
enough to expose it via a REST API. Some libraries expose and "objects" in that
format but not "classes."

This library allows us to convert PHP types into schemas and cast values based
on types. Because this library is generic enough to deal with any PHP type, we
built it completely decoupled from other systems.

Acked-by: Rodrigo Cinto <rodrigo.cinto@instapro.com>
Co-authored-by: Hugo FOUGERES <hugo.fougeres@instapro.com>