Create a converter that handles arrays as JSON
Creating a schema from arrays is very complicated because PHP doesn't have
native support for array type-hinting.

There are ways to make it fancier, using `phpdocumentor/reflection-docblock`
for example, which would help parse the doc blocks of arrays and identify the
array value type. Still, because the docblocks are simply strings, I would need
to do lots of work to get the use statements declared at the top of each file.
The `nikic/php-parser` would help with that work, but I think it's too much for
what we need right now.

This commit will introduce a converter that will handle arrays as JSON.

Acked-by: Hugo FOUGERES <hugo.fougeres@instapro.com>
Acked-by: Rodrigo Cinto <rodrigo.cinto@instapro.com>